requirements.txt
amqp==5.3.1
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
autopep8==2.3.2
billiard==4.2.1
black==24.10.0
bleach==6.2.0
boto3==1.36.6
botocore==1.36.6
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.6.10
cron-descriptor==1.4.5
cryptography==44.0.0
decorator==5.1.1
Django==5.0.6
django-admin-extra-buttons==1.5.8
django-admin-rangefilter==0.13.0
django-anymail==12.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-constance==3.1.0
django-cors-headers==4.3.1
django-debug-toolbar==4.4.2
django-environ==0.11.2
django-extensions==3.2.3
django-health-check==3.18.3
django-picklefield==3.2
django-prometheus==2.3.1
django-redis==5.4.0
django-ses==4.3.2
django-silk==5.3.2
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environ==1.0
exceptiongroup==1.2.2
executing==2.1.0
factory-boy==3.3.0
Faker==26.0.0
flower==2.0.1
gprof2dot==2024.6.6
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.4
humanize==4.11.0
idna==3.10
inflection==0.5.1
ipython==8.31.0
itypes==1.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.4.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
msgpack==1.1.0
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
s3transfer==0.11.2
sentry-sdk==2.20.0
six==1.16.0
sniffio==1.3.1
South==1.0.2
sqlparse==0.5.0
stack-data==0.6.3
setuptools==75.0.0
tinycss2==1.4.0
tomli==2.2.1
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.1
whitenoise==6.8.2
